requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
deepl==1.22.0
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.7.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
httpcore==0.9.1
huggingface-hub==0.33.4
hyperframe==5.2.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
mss==10.0.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
pynput==1.8.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-bidi==0.6.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
termcolor==3.1.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
xlib==0.21
